+++ cmd/azure/fnGetIpAddresses.go
@@ -872,3 +872,6 @@ func IncrementIp(ip net.IP) {
 		}
 	}
 }
+
+//
+//